Milliseconds timings
The current behaviour of the --show-timings flag currently only shows things in seconds. Import Linter has got much faster, it would be helpful to show it in milliseconds now, where appropriate.

This commit will show the timings in seconds with up to 3 decimals if less than 10 seconds, and in milliseconds if less than 1 second